Add IP address script for SYN-OS
This commit adds a Bash script that uses the `ip` command to retrieve and
continuously output the IP addresses and interface names for all network
interfaces on a system running SYN-OS. This script will be useful for
maintaining connectivity and checking addresses as a feed, as it does not
require additional networking software overhead and provides a simple way to
monitor network changes and troubleshoot network issues. By using the built-in
capabilities of the system, we can simplify the system and reduce the need for
additional software sources.